Modifications in the multiplication templates.

+++ csx/mult_templates/block_col_tmpl.c
@@ -5,9 +5,9 @@ static inline double block_col_${r}x${c}_case(uint8_t **ctl, uint8_t size,
 	register double *y_curr_ = *y_curr;
 	register double *x_curr_ = *x_curr;
 
-	for (uint8_t i = 0; i < ${r}; i++) {
+	for (uint64_t i = 0; i < ${r}; i++) {
 		register double yr = 0;
-		for (uint8_t j = 0; j < ${c}; j++) {
+		for (uint64_t j = 0; j < ${c}; j++) {
 			yr += x_curr_[j] * (**values);
 			(*values)++;
 		}
